// src/data/interfaces/ICarFiltersInterface.d.ts
export interface ICarFiltersInterface {
  bodyType?: string | string[];
  vehicleType?: string | string[];
  fuelType?: string | string[];
  transmissionType?: string;
  minPrice?: number;
  phone?: string;
  maxPrice?: number;
  minMileage?: number;
  maxMileage?: number;
  minYear?: number;
  maxYear?: number;
  color?: string | string[];
  painting?: string | string[];
  name?: string | string[];
  brand?: string;
  maxEngineDisplacement?: number;
  minEngineDisplacement?: number;
  minPowerkw?: number;
  maxPowerkw?: number;
  minPowerhv?: number;
  maxPowerhv?: number;
  minTorque?: number;
  maxTorque?: number;
  minTowingWeight?: number;
  maxTowingWeight?: number;
  minTotalWeight?: number;
  maxTotalWeight?: number;
  numberOfPeople?: number;
  numberOfDoors?: number;
  minAcceleration?: number;
  maxAcceleration?: number;
  minCo2Emissions?: number;
  maxCo2Emissions?: number;
  antiLockBrakingSystem: false;
  hillStartAssist: false;
  parkingAssistance: false;
  fogLights: false;
  airbags: false;
  collisionWarning: false;
  theftAlarm: false;
  xenonHeadlights: false;
  adaptiveHeadlights: false;
  electronicStabilityControl: false;
  emergencyBrakeAssist: false;
  isofixReady: false;
  corneringLights: false;
  blindSpotAssist: false;
  engineImmobilizer: false;
  ledHeadlights: false;
  tractionControlSystem: false;
  seatHeaters: false;
  rearViewCamera: false;
  fuelBatteryOperatedAuxiliaryHeater: false;
  parkingSensors: false;
  tirePressureMonitoringSystem: false;
  internalSocket: false;
  startStopSystem: false;
  skiHatch: false;
  powerWindows: false;
  electricMirrors: false;
  electricallyAdjustableSeats: false;
  electricallyOperatedTailgate: false;
  luggageCompartmentSafetyNet: false;
  sportSeats: false;
  cruiseControl: false;
  onBoardComputer: false;
  keyless: false;
  airConditionedSeating: false;
  airConditioning: false;
  retractableRearSeats: false;
  laneDepartureWarning: false;
  roofOpeningMechanism: false;
  centralLocking: false;
  motorHeater: false;
  heatedSteeringWheel: false;
  heatedWindscreen: false;
  leatherInterior: false;
  satelliteNavigator: false;
  speedLimitSensor: false;
  powerSteering: false;
  usb: false;
  smartphoneIntegration: false;
  audioSystem: false;
  audioConnection: false;
  bluetooth: false;
  headUpDisplay: false;
  mobilePhoneEquipment: false;
  alloyWheels: false;
  sunroof: false;
  multifunctionSteeringWheel: false;
  rainSensor: false;
  trunkTemperatureControl: false;
  turbo: false;
  sportsSuspension: false;
  towBar: false;
  batteryPreHeating: false;
  chassisKit: false;
  serviceBook: false;
  plugInChargingSocketForHybridCar: false;
  airSuspension: false;
  handicapEquipment: false;
  twoSetsOfTires: false;
  roofHatch: false;
}